OS: Added "swagselectionmethod", "swagdeletionmethod", "swagconfirmationmethod" and "buttonclickmethod" to valid s$strings list.
OH: "swagselectionmethod", "swagdeletionmethod", "swagconfirmationmethod" and "buttonclickmethod" are now settings that are drawn from the table map file. The settings that might be saved in the registry from previous OpenHoldem versions are deprecated and are now ignored for these methods.  SWAG timings are still set in OpenHoldem preferences, as usual. If these methods are not specified in the .tm file, then the default actions are:
  swagselectionmethod = Dbl Click
  swagdeletionmethod = Delete
  swagconfirmationmethod = Enter
  buttonclickmethod = Double
- Added new Table Map "swagselectionmethod" option: Single Click

git-svn-id: https://openholdembot.googlecode.com/svn/trunk@586 b0b9fd3a-3a3c-0410-a9aa-dfaef611493d